DDCNL-9071: Refactor
