+++ app/uk/gov/hmrc/tai/model/domain/income/Incomes.scala
@@ -16,8 +16,10 @@
 
 package uk.gov.hmrc.tai.model.domain.income
 
-import play.api.libs.json.{Format, Json}
-import uk.gov.hmrc.tai.model.domain.NonTaxCodeIncomeComponentType
+import play.api.libs.json.{Format, JsResult, JsSuccess, JsValue, Json, Reads}
+import uk.gov.hmrc.tai.model.domain.{NonTaxCodeIncomeComponentType, NpsIabdSummary}
+import uk.gov.hmrc.tai.model.domain.NpsIabdSummary.iabdsFromTotalLiabilityReads
+import uk.gov.hmrc.tai.model.domain._
 
 case class UntaxedInterest(
   incomeComponentType: NonTaxCodeIncomeComponentType,
@@ -39,6 +41,57 @@ case class OtherNonTaxCodeIncome(
 
 object OtherNonTaxCodeIncome {
   implicit val format: Format[OtherNonTaxCodeIncome] = Json.format[OtherNonTaxCodeIncome]
+
+  val nonTaxCodeIncomeReads = new Reads[Seq[OtherNonTaxCodeIncome]] {
+    override def reads(json: JsValue): JsResult[Seq[OtherNonTaxCodeIncome]] = {
+      val extractedIabds: Seq[NpsIabdSummary] = json.as[Seq[NpsIabdSummary]](iabdsFromTotalLiabilityReads)
+      JsSuccess(nonTaxCodeIncomes(extractedIabds))
+    }
+  }
+
+  private def nonTaxCodeIncomes(iabds: Seq[NpsIabdSummary]): Seq[OtherNonTaxCodeIncome] =
+    iabds collect {
+      case iabd if nonTaxCodeIncomesMap.isDefinedAt(iabd.componentType) =>
+        OtherNonTaxCodeIncome(
+          nonTaxCodeIncomesMap(iabd.componentType),
+          iabd.employmentId,
+          iabd.amount,
+          iabd.description
+        )
+    }
+
+  private val nonTaxCodeIncomesMap: Map[Int, NonTaxCodeIncomeComponentType] = Map(
+    19  -> NonCodedIncome,
+    20  -> Commission,
+    21  -> OtherIncomeEarned,
+    22  -> OtherIncomeNotEarned,
+    23  -> PartTimeEarnings,
+    24  -> Tips,
+    25  -> OtherEarnings,
+    26  -> CasualEarnings,
+    62  -> ForeignDividendIncome,
+    63  -> ForeignPropertyIncome,
+    64  -> ForeignInterestAndOtherSavings,
+    65  -> ForeignPensionsAndOtherIncome,
+    66  -> StatePension,
+    67  -> OccupationalPension,
+    68  -> PublicServicesPension,
+    69  -> ForcesPension,
+    70  -> PersonalPensionAnnuity,
+    72  -> Profit,
+    75  -> BankOrBuildingSocietyInterest,
+    76  -> UkDividend,
+    77  -> UnitTrust,
+    78  -> StockDividend,
+    79  -> NationalSavings,
+    80  -> SavingsBond,
+    81  -> PurchasedLifeAnnuities,
+    82  -> UntaxedInterestIncome,
+    83  -> IncapacityBenefit,
+    84  -> JobSeekersAllowance,
+    123 -> EmploymentAndSupportAllowance
+  )
+
 }
 
 case class NonTaxCodeIncome(
