+++ app/uk/gov/hmrc/tai/model/domain/income/TaxCodeIncome.scala
@@ -19,8 +19,8 @@ package uk.gov.hmrc.tai.model.domain.income
 import java.time.LocalDate
 import play.api.Logger
 import play.api.libs.json._
+import uk.gov.hmrc.tai.model.domain.IabdSummary.iabdSummaryReads
 import uk.gov.hmrc.tai.model.domain._
-import uk.gov.hmrc.tai.model.domain.formatters.income.TaxCodeIncomeHodFormatters
 import uk.gov.hmrc.tai.util.{TaiConstants, TaxCodeHistoryConstants}
 
 sealed trait BasisOperation
@@ -123,7 +123,7 @@ case class TaxCodeIncome(
   }
 }
 
-object TaxCodeIncome extends TaxCodeIncomeHodFormatters {
+object TaxCodeIncome {
 
   implicit val writes: Writes[TaxCodeIncome] = new Writes[TaxCodeIncome] {
     override def writes(o: TaxCodeIncome): JsValue =
@@ -149,5 +149,116 @@ object TaxCodeIncome extends TaxCodeIncomeHodFormatters {
         }
       )
   }
+
+  private val logger: Logger = Logger(getClass.getName)
+
+  private val basisOperationReads = new Reads[BasisOperation] {
+    override def reads(json: JsValue): JsResult[BasisOperation] = {
+      val result = json.asOpt[Int] match {
+        case Some(1) => Week1Month1BasisOperation
+        case _       => OtherBasisOperation
+      }
+      JsSuccess(result)
+    }
+  }
+
+  val taxCodeIncomeSourceReads = new Reads[TaxCodeIncome] {
+    override def reads(json: JsValue) = {
+      val incomeSourceType = taxCodeIncomeType(json)
+      val employmentId = (json \ "employmentId").asOpt[Int]
+      val amount = totalTaxableIncome(json, employmentId).getOrElse(BigDecimal(0))
+      val description = incomeSourceType.toString
+      val taxCode = (json \ "taxCode").asOpt[String].getOrElse("")
+      val name = (json \ "name").asOpt[String].getOrElse("")
+      val basisOperation =
+        (json \ "basisOperation").asOpt[BasisOperation](basisOperationReads).getOrElse(OtherBasisOperation)
+      val status = employmentStatus(json)
+      val iyaCy = (json \ "inYearAdjustmentIntoCY").asOpt[BigDecimal].getOrElse(BigDecimal(0))
+      val totalIya = (json \ "totalInYearAdjustment").asOpt[BigDecimal].getOrElse(BigDecimal(0))
+      val iyaCyPlusOne = (json \ "inYearAdjustmentIntoCYPlusOne").asOpt[BigDecimal].getOrElse(BigDecimal(0))
+      JsSuccess(
+        TaxCodeIncome(
+          incomeSourceType,
+          employmentId,
+          amount,
+          description,
+          taxCode,
+          name,
+          basisOperation,
+          status,
+          iyaCy,
+          totalIya,
+          iyaCyPlusOne
+        )
+      )
+    }
+  }
+
+  val taxCodeIncomeSourcesReads = new Reads[Seq[TaxCodeIncome]] {
+    override def reads(json: JsValue): JsResult[Seq[TaxCodeIncome]] = {
+      val taxCodeIncomes = (json \ "incomeSources")
+        .asOpt[Seq[TaxCodeIncome]](Reads.seq(taxCodeIncomeSourceReads))
+        .getOrElse(Seq.empty[TaxCodeIncome])
+      JsSuccess(taxCodeIncomes)
+    }
+  }
+
+  private def taxCodeIncomeType(json: JsValue): TaxComponentType = {
+    def indicator(indicatorType: String) = (json \ indicatorType).asOpt[Boolean].getOrElse(false)
+
+    if (indicator("pensionIndicator"))
+      PensionIncome
+    else if (indicator("jsaIndicator"))
+      JobSeekerAllowanceIncome
+    else if (indicator("otherIncomeSourceIndicator"))
+      OtherIncome
+    else
+      EmploymentIncome
+  }
+
+  private def employmentStatus(json: JsValue): TaxCodeIncomeStatus = {
+    val employmentStatus = (json \ "employmentStatus").asOpt[Int]
+
+    employmentStatus match {
+      case Some(1) => Live
+      case Some(2) => PotentiallyCeased
+      case Some(3) => Ceased
+      case default =>
+        logger.warn(s"Invalid Employment Status -> $default")
+        throw new RuntimeException("Invalid employment status")
+    }
+  }
+
+  private def totalTaxableIncome(json: JsValue, employmentId: Option[Int]): Option[BigDecimal] = {
+    val iabdSummaries: Option[Seq[IabdSummary]] =
+      (json \ "payAndTax" \ "totalIncome" \ "iabdSummaries").asOpt[Seq[IabdSummary]](Reads.seq(iabdSummaryReads))
+    val iabdSummary = iabdSummaries.flatMap {
+      _.find(iabd =>
+        newEstimatedPayTypeFilter(iabd) &&
+          employmentFilter(iabd, employmentId)
+      )
+    }
+    iabdSummary.map(_.amount) match {
+      case Some(amount) => Some(amount)
+      case _ =>
+        logger.warn("TotalTaxableIncome is 0")
+        None
+    }
+  }
+
+  def newEstimatedPayTypeFilter(iabdSummary: IabdSummary): Boolean = {
+    val NewEstimatedPay = 27
+    iabdSummary.componentType == NewEstimatedPay
+  }
+
+  def employmentFilter(iabdSummary: IabdSummary, employmentId: Option[Int]): Boolean = {
+    val compare = for {
+      jsEmploymentId <- iabdSummary.employmentId
+      empId          <- employmentId
+    } yield jsEmploymentId == empId
+
+    compare.getOrElse(false)
+  }
+
   implicit val reads: Reads[TaxCodeIncome] = taxCodeIncomeSourceReads
 }
