+++ app/uk/gov/hmrc/tai/model/domain/taxAdjustments/TaxAdjustmentComponent.scala
@@ -90,6 +90,120 @@ case class TaxAdjustmentComponent(taxAdjustmentType: TaxAdjustmentType, taxAdjus
 
 object TaxAdjustmentComponent {
   implicit val format: Format[TaxAdjustmentComponent] = Json.format[TaxAdjustmentComponent]
+
+  def flattenTaxAdjustmentComponents(components: Option[TaxAdjustmentComponent]*): Seq[TaxAdjustmentComponent] =
+    components.toSeq.flatten
+
+  def readTaxAdjustmentComponent(
+    json: JsValue,
+    taxAdjustmentTypeInJson: String,
+    taxAdjustmentType: TaxAdjustmentType
+  ): Option[TaxAdjustmentComponent] =
+    (json \ taxAdjustmentTypeInJson).asOpt[BigDecimal].collect { case amount =>
+      TaxAdjustmentComponent(taxAdjustmentType, amount)
+    }
+
+  val taxReliefFormattersReads = new Reads[Seq[TaxAdjustmentComponent]] {
+    override def reads(json: JsValue): JsResult[Seq[TaxAdjustmentComponent]] =
+      (json \ "totalLiability" \ "basicRateExtensions").asOpt[JsObject] match {
+        case Some(js) =>
+          val personalPensionPayment = readTaxAdjustmentComponent(js, "personalPensionPayment", PersonalPensionPayment)
+          val personalPensionPaymentRelief =
+            readTaxAdjustmentComponent(js, "personalPensionPaymentRelief", PersonalPensionPaymentRelief)
+          val giftAidPaymentsRelief = readTaxAdjustmentComponent(js, "giftAidPaymentsRelief", GiftAidPaymentsRelief)
+          val giftAidPayments = readTaxAdjustmentComponent(js, "giftAidPayments", GiftAidPayments)
+
+          JsSuccess(
+            flattenTaxAdjustmentComponents(
+              personalPensionPayment,
+              personalPensionPaymentRelief,
+              giftAidPaymentsRelief,
+              giftAidPayments
+            )
+          )
+        case _ => JsSuccess(Seq.empty[TaxAdjustmentComponent])
+      }
+  }
+
+  val alreadyTaxedAtSourceReads = new Reads[Seq[TaxAdjustmentComponent]] {
+    override def reads(json: JsValue): JsResult[Seq[TaxAdjustmentComponent]] =
+      (json \ "totalLiability" \ "alreadyTaxedAtSource").asOpt[JsObject] match {
+        case Some(js) =>
+          val taxOnBankInterest = readTaxAdjustmentComponent(js, "taxOnBankBSInterest", TaxOnBankBSInterest)
+          val taxOnUkDividends = readTaxAdjustmentComponent(js, "taxCreditOnUKDividends", TaxCreditOnUKDividends)
+          val taxOnForeignInterest =
+            readTaxAdjustmentComponent(js, "taxCreditOnForeignInterest", TaxCreditOnForeignInterest)
+          val taxOnForeignDividends =
+            readTaxAdjustmentComponent(js, "taxCreditOnForeignIncomeDividends", TaxCreditOnForeignIncomeDividends)
+
+          JsSuccess(
+            flattenTaxAdjustmentComponents(
+              taxOnBankInterest,
+              taxOnUkDividends,
+              taxOnForeignInterest,
+              taxOnForeignDividends
+            )
+          )
+        case _ => JsSuccess(Seq.empty[TaxAdjustmentComponent])
+      }
+  }
+
+  val otherTaxDueReads = new Reads[Seq[TaxAdjustmentComponent]] {
+    override def reads(json: JsValue): JsResult[Seq[TaxAdjustmentComponent]] =
+      (json \ "totalLiability" \ "otherTaxDue").asOpt[JsObject] match {
+        case Some(js) =>
+          val excessGiftAidTax = readTaxAdjustmentComponent(js, "excessGiftAidTax", ExcessGiftAidTax)
+          val excessWidowsAndOrphans = readTaxAdjustmentComponent(js, "excessWidowsAndOrphans", ExcessWidowsAndOrphans)
+          val pensionPaymentsAdjustment =
+            readTaxAdjustmentComponent(js, "pensionPaymentsAdjustment", PensionPaymentsAdjustment)
+          val childBenefit = readTaxAdjustmentComponent(js, "childBenefit", ChildBenefit)
+
+          JsSuccess(
+            flattenTaxAdjustmentComponents(
+              excessGiftAidTax,
+              excessWidowsAndOrphans,
+              pensionPaymentsAdjustment,
+              childBenefit
+            )
+          )
+        case _ => JsSuccess(Seq.empty[TaxAdjustmentComponent])
+      }
+  }
+
+  val reliefsGivingBackTaxReads = new Reads[Seq[TaxAdjustmentComponent]] {
+    override def reads(json: JsValue): JsResult[Seq[TaxAdjustmentComponent]] =
+      (json \ "totalLiability" \ "reliefsGivingBackTax").asOpt[JsObject] match {
+        case Some(js) =>
+          val enterpriseInvestment =
+            readTaxAdjustmentComponent(js, "enterpriseInvestmentSchemeRelief", EnterpriseInvestmentSchemeRelief)
+          val concession = readTaxAdjustmentComponent(js, "concessionalRelief", ConcessionalRelief)
+          val maintenancePayments = readTaxAdjustmentComponent(js, "maintenancePayments", MaintenancePayments)
+          val marriedCouplesAllowance =
+            readTaxAdjustmentComponent(js, "marriedCouplesAllowance", MarriedCouplesAllowance)
+          val doubleTaxation = readTaxAdjustmentComponent(js, "doubleTaxationRelief", DoubleTaxationRelief)
+
+          JsSuccess(
+            flattenTaxAdjustmentComponents(
+              enterpriseInvestment,
+              concession,
+              maintenancePayments,
+              marriedCouplesAllowance,
+              doubleTaxation
+            )
+          )
+        case _ => JsSuccess(Seq.empty[TaxAdjustmentComponent])
+      }
+  }
+
+  val taxAdjustmentComponentReads = new Reads[Seq[TaxAdjustmentComponent]] {
+    override def reads(json: JsValue): JsResult[Seq[TaxAdjustmentComponent]] = {
+      val reliefsGivingBackComponents = json.as[Seq[TaxAdjustmentComponent]](reliefsGivingBackTaxReads)
+      val otherTaxDues = json.as[Seq[TaxAdjustmentComponent]](otherTaxDueReads)
+      val alreadyTaxedAtSources = json.as[Seq[TaxAdjustmentComponent]](alreadyTaxedAtSourceReads)
+      val taxReliefComponent = json.as[Seq[TaxAdjustmentComponent]](taxReliefFormattersReads)
+      JsSuccess(reliefsGivingBackComponents ++ otherTaxDues ++ alreadyTaxedAtSources ++ taxReliefComponent)
+    }
+  }
 }
 
 case class TaxAdjustment(amount: BigDecimal, taxAdjustmentComponents: Seq[TaxAdjustmentComponent])
