+++ app/uk/gov/hmrc/tai/service/CodingComponentService.scala
@@ -21,14 +21,13 @@ import uk.gov.hmrc.domain.Nino
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.tai.connectors.TaxAccountConnector
 import uk.gov.hmrc.tai.model.domain.calculation.CodingComponent
-import uk.gov.hmrc.tai.model.domain.formatters.taxComponents.TaxAccountHodFormatters
+import uk.gov.hmrc.tai.model.domain.calculation.CodingComponent.codingComponentReads
 import uk.gov.hmrc.tai.model.tai.TaxYear
 
 import scala.concurrent.{ExecutionContext, Future}
 
 @Singleton
-class CodingComponentService @Inject() (taxAccountConnector: TaxAccountConnector)(implicit ec: ExecutionContext)
-    extends TaxAccountHodFormatters {
+class CodingComponentService @Inject() (taxAccountConnector: TaxAccountConnector)(implicit ec: ExecutionContext) {
 
   def codingComponents(nino: Nino, year: TaxYear)(implicit hc: HeaderCarrier): Future[Seq[CodingComponent]] =
     taxAccountConnector.taxAccount(nino, year).map(_.as[Seq[CodingComponent]](codingComponentReads))
