+++ app/uk/gov/hmrc/tai/service/TaxAccountSummaryService.scala
@@ -21,9 +21,7 @@ import play.api.mvc.Request
 import uk.gov.hmrc.domain.Nino
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.tai.model.domain.calculation.CodingComponent
-import uk.gov.hmrc.tai.model.domain.formatters.TaxAccountSummaryHodFormatters
 import uk.gov.hmrc.tai.model.domain._
-import uk.gov.hmrc.tai.model.domain.formatters.taxComponents.TaxAccountHodFormatters
 import uk.gov.hmrc.tai.model.tai.TaxYear
 import uk.gov.hmrc.tai.service.helper.TaxAccountHelper
 
@@ -38,7 +36,7 @@ class TaxAccountSummaryService @Inject() (
   taxAccountHelper: TaxAccountHelper
 )(implicit
   ec: ExecutionContext
-) extends TaxAccountSummaryHodFormatters with TaxAccountHodFormatters {
+) {
 
   def taxAccountSummary(nino: Nino, year: TaxYear)(implicit
     hc: HeaderCarrier,
