+++ test/uk/gov/hmrc/tai/repositories/IncomeRepositorySpec.scala
@@ -21,7 +21,6 @@ import org.mockito.ArgumentMatchers.{any, eq => meq}
 import play.api.libs.json._
 import uk.gov.hmrc.tai.connectors.TaxAccountConnector
 import uk.gov.hmrc.tai.model.domain._
-import uk.gov.hmrc.tai.model.domain.formatters.IabdDetails
 import uk.gov.hmrc.tai.model.domain.income._
 import uk.gov.hmrc.tai.model.tai.TaxYear
 import uk.gov.hmrc.tai.repositories.deprecated.IncomeRepository
